Tidy up generation code, ensure scalars are passed as such from Delphi
