+++ cpp_generation/src/utils/config.py
@@ -76,19 +76,6 @@ def get_c_write_data_single_year(config_name, name, cfg):
     return f'write_data<{cfg["num_type"]}, {dim_len(cfg)}>(state.{name}, out.{config_name}->{name}, out.{config_name}->{name}_length, "{name}");'
 
 
-def get_delphi_setter(name, cfg):
-  name = to_camel_case(name)
-  return f'Set{ name }(const in{ name }: TGBFixedArray<{get_delphi_num_type(cfg)}>);'
-
-
-def get_delphi_param_setter(config_name):
-  return f'Set{ config_name }Params(const { config_name.lower() }Params: Leapfrog{ config_name }ParamsView);'
-
-
-def get_delphi_state_setter(config_name):
-  return f'Set{ config_name }State(const { config_name.lower() }State: Leapfrog{ config_name }StateView);'
-
-
 def get_reset(name, cfg):
   if cfg["type"] == "scalar":
     return f'{name} = {cfg.get("default") or 0}'
@@ -110,36 +97,3 @@ def get_output_state_chip(name, cfg):
 
 def get_num_type(cfg):
   return cfg["num_type"]
-
-
-def get_num_type_delphi(cfg):
-  n_dim = dim_len(cfg)
-  if n_dim == 0:
-    return cfg["num_type"]
-  else:
-    return f'{cfg["num_type"]}*'
-
-def dict_len(dict):
-  return len(dict.keys())
-
-
-def to_camel_case(snake_str):
-  return "".join(x.capitalize() for x in snake_str.lower().split("_"))
-
-
-def to_lower_camel_case(snake_str):
-  camel_string = to_camel_case(snake_str)
-  return snake_str[0].lower() + camel_string[1:]
-
-
-def get_delphi_ptr_type(cfg):
-  return 'P' + get_delphi_num_type(cfg)
-
-
-def get_delphi_num_type(cfg):
-  if cfg["num_type"] == 'real_type':
-    return 'Double'
-  elif cfg["num_type"] == 'int':
-    return 'Integer'
-  else:
-    raise Exception(f'Unknown type {cfg["num_type"]}, must be one of "real_type" or "int".')
