+++ delphi/LeapfrogInterface.pas
@@ -89,7 +89,7 @@ type
     cd4ProgressionLength: Integer;
     cd4InitialDistribution: PDouble;
     cd4InitialDistributionLength: Integer;
-    scaleCd4Mortality: PInteger;
+    scaleCd4Mortality: Integer;
     scaleCd4MortalityLength: Integer;
     idxHmElig: PInteger;
     idxHmEligLength: Integer;
@@ -97,7 +97,7 @@ type
     mortalityLength: Integer;
     mortalityTimeRateRatio: PDouble;
     mortalityTimeRateRatioLength: Integer;
-    dropoutRecoverCd4: PInteger;
+    dropoutRecoverCd4: Integer;
     dropoutRecoverCd4Length: Integer;
     dropoutRate: PDouble;
     dropoutRateLength: Integer;
@@ -105,7 +105,7 @@ type
     adultsOnArtLength: Integer;
     adultsOnArtIsPercent: PInteger;
     adultsOnArtIsPercentLength: Integer;
-    initiationMortalityWeight: PDouble;
+    initiationMortalityWeight: Double;
     initiationMortalityWeightLength: Integer;
     hArtStageDur: PDouble;
     hArtStageDurLength: Integer;
@@ -120,15 +120,15 @@ type
     cd4Mortality: TGBFixedArray<Double>;
     cd4Progression: TGBFixedArray<Double>;
     cd4InitialDistribution: TGBFixedArray<Double>;
-    scaleCd4Mortality: TGBFixedArray<Integer>;
+    scaleCd4Mortality: Integer;
     idxHmElig: TGBFixedArray<Integer>;
     mortality: TGBFixedArray<Double>;
     mortalityTimeRateRatio: TGBFixedArray<Double>;
-    dropoutRecoverCd4: TGBFixedArray<Integer>;
+    dropoutRecoverCd4: Integer;
     dropoutRate: TGBFixedArray<Double>;
     adultsOnArt: TGBFixedArray<Double>;
     adultsOnArtIsPercent: TGBFixedArray<Integer>;
-    initiationMortalityWeight: TGBFixedArray<Double>;
+    initiationMortalityWeight: Double;
     hArtStageDur: TGBFixedArray<Double>;
     function getView(): LeapfrogHivAdultParamsView;
     Destructor Destroy; override;
@@ -266,9 +266,9 @@ type
     totalBirthsLength: Integer;
     ctxEffect: PDouble;
     ctxEffectLength: Integer;
-    hcArtStart: PDouble;
+    hcArtStart: Double;
     hcArtStartLength: Integer;
-    localAdjFactor: PDouble;
+    localAdjFactor: Double;
     localAdjFactorLength: Integer;
 end;
 
@@ -319,8 +319,8 @@ type
     adultFemaleHivnpop: TGBFixedArray<Double>;
     totalBirths: TGBFixedArray<Double>;
     ctxEffect: TGBFixedArray<Double>;
-    hcArtStart: TGBFixedArray<Double>;
-    localAdjFactor: TGBFixedArray<Double>;
+    hcArtStart: Double;
+    localAdjFactor: Double;
     function getView(): LeapfrogHivChildParamsView;
     Destructor Destroy; override;
 end;
@@ -464,15 +464,12 @@ begin;
   cd4Mortality.Free;
   cd4Progression.Free;
   cd4InitialDistribution.Free;
-  scaleCd4Mortality.Free;
   idxHmElig.Free;
   mortality.Free;
   mortalityTimeRateRatio.Free;
-  dropoutRecoverCd4.Free;
   dropoutRate.Free;
   adultsOnArt.Free;
   adultsOnArtIsPercent.Free;
-  initiationMortalityWeight.Free;
   hArtStageDur.Free;
   inherited;
 end;
@@ -505,24 +502,24 @@ begin;
   Result.cd4ProgressionLength := cd4Progression.GetLength();
   Result.cd4InitialDistribution := PDouble(cd4InitialDistribution.data);
   Result.cd4InitialDistributionLength := cd4InitialDistribution.GetLength();
-  Result.scaleCd4Mortality := PInteger(scaleCd4Mortality.data);
-  Result.scaleCd4MortalityLength := scaleCd4Mortality.GetLength();
+  Result.scaleCd4Mortality := scaleCd4Mortality;
+  Result.scaleCd4MortalityLength := 1;
   Result.idxHmElig := PInteger(idxHmElig.data);
   Result.idxHmEligLength := idxHmElig.GetLength();
   Result.mortality := PDouble(mortality.data);
   Result.mortalityLength := mortality.GetLength();
   Result.mortalityTimeRateRatio := PDouble(mortalityTimeRateRatio.data);
   Result.mortalityTimeRateRatioLength := mortalityTimeRateRatio.GetLength();
-  Result.dropoutRecoverCd4 := PInteger(dropoutRecoverCd4.data);
-  Result.dropoutRecoverCd4Length := dropoutRecoverCd4.GetLength();
+  Result.dropoutRecoverCd4 := dropoutRecoverCd4;
+  Result.dropoutRecoverCd4Length := 1;
   Result.dropoutRate := PDouble(dropoutRate.data);
   Result.dropoutRateLength := dropoutRate.GetLength();
   Result.adultsOnArt := PDouble(adultsOnArt.data);
   Result.adultsOnArtLength := adultsOnArt.GetLength();
   Result.adultsOnArtIsPercent := PInteger(adultsOnArtIsPercent.data);
   Result.adultsOnArtIsPercentLength := adultsOnArtIsPercent.GetLength();
-  Result.initiationMortalityWeight := PDouble(initiationMortalityWeight.data);
-  Result.initiationMortalityWeightLength := initiationMortalityWeight.GetLength();
+  Result.initiationMortalityWeight := initiationMortalityWeight;
+  Result.initiationMortalityWeightLength := 1;
   Result.hArtStageDur := PDouble(hArtStageDur.data);
   Result.hArtStageDurLength := hArtStageDur.GetLength();
   inherited;
@@ -597,8 +594,6 @@ begin;
   adultFemaleHivnpop.Free;
   totalBirths.Free;
   ctxEffect.Free;
-  hcArtStart.Free;
-  localAdjFactor.Free;
   inherited;
 end;
 
@@ -710,10 +705,10 @@ begin;
   Result.totalBirthsLength := totalBirths.GetLength();
   Result.ctxEffect := PDouble(ctxEffect.data);
   Result.ctxEffectLength := ctxEffect.GetLength();
-  Result.hcArtStart := PDouble(hcArtStart.data);
-  Result.hcArtStartLength := hcArtStart.GetLength();
-  Result.localAdjFactor := PDouble(localAdjFactor.data);
-  Result.localAdjFactorLength := localAdjFactor.GetLength();
+  Result.hcArtStart := hcArtStart;
+  Result.hcArtStartLength := 1;
+  Result.localAdjFactor := localAdjFactor;
+  Result.localAdjFactorLength := 1;
   inherited;
 end;
